csv_to_pkl.py converts CSV data from norgate 2D numpy arrays and writes them out as pickled objects. One file per stock.
DataAccess.py no longer supports HDF. It reads in pickled data created by csv_to_pkl.py and returns a pandas object containing the data asked for. The main function in this file serves to demonstrate its use.

The DataAccessExample.py and GekkoDataAccessTester.py are no longer needed because they were examples of the old DataAccess that was based on HDF. Hence deleting.


Former-commit-id: 95de7dcb690e9ffe3bf5f5290d1b593b4a212482
Former-commit-id: 095c66d02413fcaf9b1187cad68f4a070dbcda44