csv_to_pkl.py converts CSV data from norgate 2D numpy arrays and writes them out as pickled objects. One file per stock.
DataAccess.py no longer supports HDF. It reads in pickled data created by csv_to_pkl.py and returns a pandas object containing the data asked for. The main function in this file serves to demonstrate its use.

The DataAccessExample.py and GekkoDataAccessTester.py are no longer needed because they were examples of the old DataAccess that was based on HDF. Hence deleting.


Former-commit-id: 6fa98e16e0a3387ad6b43ac00a426730b7c7bf6c
Former-commit-id: ef62381734355a8a950e42338a01f8b97941fd59